+++ src/Forms/UI/views/layouts/window.blade.php
@@ -1,0 +1,7 @@
+@if (count($getComponents()) > 0)
+    <x-chief::window :title="$getTitle()" :description="$getDescription()" :variant="$getVariant()">
+        @foreach ($getComponents() as $childComponent)
+            {{ $childComponent }}
+        @endforeach
+    </x-chief::window>
+@endif